Make it compatible with python 3.10 and 3.11 on Windows

+++ fast_align_audio/_alignment_cffi.py
@@ -5,7 +5,7 @@ from cffi import FFI
 ffibuilder = FFI()
 ffibuilder.cdef("""
     typedef struct {
-        ssize_t min_idx;
+        int min_idx;
         float min_val;
     } MinResult;
 """)
